+++ pytest_container/runtime.py
@@ -200,10 +200,6 @@ class OciRuntimeABC(ABC):
 
         """
 
-    @abstractmethod
-    def get_image_id_from_stdout(self, stdout: str) -> str:
-        """Returns the image id/hash from the stdout of a build command."""
-
     def get_container_health(self, container_id: str) -> ContainerHealth:
         """Inspects the running container with the supplied id and returns its current
         health.
@@ -459,12 +455,6 @@ class PodmanRuntime(OciRuntimeBase):
             _runtime_functional=self._runtime_functional,
         )
 
-    def get_image_id_from_stdout(self, stdout: str) -> str:
-        # buildah prints the full image hash to the last non-empty line
-        return list(
-            filter(None, map(lambda l: l.strip(), stdout.split("\n")))
-        )[-1]
-
     # pragma pylint: disable=used-before-assignment
     @cached_property
     def version(self) -> Version:
@@ -550,15 +540,6 @@ class DockerRuntime(OciRuntimeBase):
             runner_binary="docker",
             _runtime_functional=self._runtime_functional,
         )
-
-    def get_image_id_from_stdout(self, stdout: str) -> str:
-        # docker build prints this into the last non-empty line:
-        # Successfully built 1e3c746e8069
-        # -> grab the last line (see podman) & the last entry
-        last_line = list(
-            filter(None, map(lambda l: l.strip(), stdout.split("\n")))
-        )[-1]
-        return last_line.split()[-1]
 
     @cached_property
     def version(self) -> Version: